Fix #170: moving to independent module
