--- camel-k-runtime-core/src/main/java/org/apache/camel/k/WebhookAction.java
@@ -1,8 +1,0 @@
-package org.apache.camel.k;
-
-/**
- * Represents an action to perform on a remote webhook endpoint.
- */
-public enum  WebhookAction {
-    REGISTER, UNREGISTER;
-}
