+++ camel-k-runtime-main/src/main/java/org/apache/camel/k/main/Application.java
@@ -36,12 +36,11 @@ public final class Application {
 
     private Application() {
     }
-    
+
     public static void main(String[] args) throws Exception {
         ApplicationRuntime runtime = new ApplicationRuntime();
         runtime.setProperties(PropertiesSupport.loadProperties());
         runtime.addListeners(ServiceLoader.load(Runtime.Listener.class));
-        runtime.getCamelContext().addRoutePolicyFactory(new WebhookRoutePolicyFactory());
         runtime.run();
     }
 }
