+++ camel-k-runtime-webhook/src/main/java/org/apache/camel/k/webhook/WebhookAction.java
@@ -1,0 +1,5 @@
+package org.apache.camel.k.webhook;
+
+public enum WebhookAction {
+    REGISTER, UNREGISTER
+}
