+++ camel-k-runtime-webhook/src/test/resources/webhook.js
